Split modelStore.ts monolith into focused state slices
The 1030-line store handled geometry, materials, BC/load groups, moment
math, results, and view state in one file (issue #202). It is now a thin
combiner over focused slices, each typed against the full ModelState so
cross-slice transitions (a new mesh clearing BC groups, etc.) stay
type-checked:

- geometrySlice.ts: nodes/elements, imported CAD geometry, meshing state
- materialSlice.ts: material definitions + CRUD
- boundarySlice.ts: BC/load groups, rebuildConstraints/Loads/SurfaceLoads,
  pick-session state
- momentLoad.ts:    pure moment -> equivalent-nodal-forces math, no Zustand
- resultsSlice.ts:  solver output, solve settings, stage transitions
- viewSlice.ts:     transient viewport presentation settings
- modelStore.ts:    combiner + whole-model lifecycle (loadAnalysis, reset),
  re-exporting every public type so consumers are untouched

momentLoad.ts is now unit-tested without any store setup
(tests/test_moment_load.mjs, wired into the bun test chain): zero net
force, exact moment reproduction, per-axis superposition, and the
on-axis-face skip warning.

Closes #202
