+++ web/package.json
@@ -15,7 +15,7 @@
     "format:check": "prettier --check src tests eslint-rules eslint.config.js \"../*.md\" \"../.github/**/*.md\"",
     "validate": "bun ../examples/validation/run.mjs",
     "examples:generate": "bun ../examples/web-examples/generate.mjs",
-    "test": "bun ../examples/validation/run.mjs && bun ../examples/validation/dof-constraint.test.mjs && bun ../examples/validation/prescribed-displacement.test.mjs && bun tests/test_wall_bracket.mjs 20.0 && bun tests/test_face_pick.mjs && playwright test",
+    "test": "bun ../examples/validation/run.mjs && bun ../examples/validation/dof-constraint.test.mjs && bun ../examples/validation/prescribed-displacement.test.mjs && bun tests/test_wall_bracket.mjs 20.0 && bun tests/test_face_pick.mjs && bun tests/test_moment_load.mjs && playwright test",
     "test:coverage": "rm -rf .nyc_output coverage && COVERAGE=1 playwright test && bun run coverage:report",
     "coverage:report": "nyc report && bun scripts/coverage-dead-code.ts",
     "test:ui": "playwright test --ui",
